+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 Changelog
 ---------
-* 0.1.0 (2013-12-30 Initial commit)
+- 0.2.1 (2014-01-03)
 
-    * added function autoSuggest()
+   * fixed bug in `quote()` for single or multipe query symbols
+
+- 0.2.0 (2011-01-03)
+
+    * added function `quote()`
+    
+- 0.1.0 (2013-12-30 Initial commit)
+
+    * added function `autoSuggest()`